# src/main/java/com/spring/demospring/api/UserApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (5.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.spring.demospring.api;

import java.util.Optional;

import javax.annotation.Generated;
import javax.validation.Valid;

import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.context.request.NativeWebRequest;

import com.spring.demospring.dto.ResponseModel;
import com.spring.demospring.dto.User;

import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiParam;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2022-02-19T18:02:28.634+01:00[Europe/Berlin]")
@Validated
@Api(value = "user", description = "the user API")
public interface UserApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /user : Create user
     *
     * @param user  (required)
     * @return response (status code 200)
     *         or error response (status code 400)
     */
    @ApiOperation(
    	value = "createUser",
        notes = "Create user",
        tags = { "user" }
    )
    @ApiResponses(value = {
    		@ApiResponse(code = 200, message = "successful operation", response = ResponseModel.class),
            @ApiResponse(code = 400, message = "error response", response =  ResponseModel.class),
            @ApiResponse(code = 404, message = "error response", response =  ResponseModel.class)
    	}
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/user",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<ResponseModel> createUser(
        @ApiParam(value = "User", required = true) @Valid @RequestBody User user
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 0, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /user/{id} : Delete user
     *
     * @param id The name that needs to be deleted (required)
     * @return response (status code 200)
     *         or error response (status code 400)
     *         or error response (status code 404)
     */
    @ApiOperation(
    	value = "deleteUser",
    	notes = "Delete user",
        tags = { "user" }
    )
    @ApiResponses(value = {
    		@ApiResponse(code = 200, message = "successful operation", response = ResponseModel.class),
            @ApiResponse(code = 400, message = "error response", response =  ResponseModel.class),
            @ApiResponse(code = 404, message = "error response", response =  ResponseModel.class)
    	}
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/user/{id}",
        produces = { "application/json" }
    )
    default ResponseEntity<ResponseModel> deleteUser(
        @ApiParam(name = "id", required = true) @PathVariable("id") String id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 0, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /user/{id} : Get user by id
     *
     * @param id The name that needs to be fetched (required)
     * @return successful operation (status code 200)
     *         or error response (status code 400)
     *         or error response (status code 404)
     */
    @ApiOperation(
    	value = "getUserByName",
    	notes = "Get user by id",
        tags = { "user" }
    )
    @ApiResponses(value = {
    		@ApiResponse(code = 200, message = "successful operation", response = User.class),
            @ApiResponse(code = 400, message = "error response", response =  ResponseModel.class),
            @ApiResponse(code = 404, message = "error response", response =  ResponseModel.class)
    	}
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/user/{id}",
        produces = { "application/json" }
    )
    default ResponseEntity<User> getUserByName(
    		@ApiParam(name = "id", required = true) @PathVariable("id") String id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"password\" : \"password\", \"address\" : \"address\", \"surname\" : \"surname\", \"name\" : \"name\", \"id\" : 0, \"email\" : \"email\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /user/{id} : Updated user
     *
     * @param id name that need to be updated (required)
     * @param user  (required)
     * @return response (status code 200)
     *         or error response (status code 400)
     *         or error response (status code 404)
     */
    @ApiOperation(
        value = "updateUser",
        notes = "Updated user",
        tags = { "user" }
    )
    @ApiResponses(value = {
    		@ApiResponse(code = 200, message = "successful operation", response = User.class),
            @ApiResponse(code = 400, message = "error response", response =  ResponseModel.class),
            @ApiResponse(code = 404, message = "error response", response =  ResponseModel.class)
    	}
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/user/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<ResponseModel> updateUser(
    	@ApiParam(name = "id", required = true) @PathVariable("id") String id,
    	@ApiParam(name = "User") @Valid @RequestBody User user
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 0, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}