# src/main/resources/openapi/Autogenerated/src/main/java/com/spring/demospring/api/SearchUserApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (5.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.spring.demospring.api;

import org.openapitools.model.ResponseModel;
import org.openapitools.model.User;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2022-02-19T18:02:28.634+01:00[Europe/Berlin]")
@Validated
@Tag(name = "searchUser", description = "the searchUser API")
public interface SearchUserApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /searchUser : Get user by parameters
     *
     * @param name The name that needs to be fetched. Use user1 for testing.  (optional)
     * @param surname The surname that needs to be fetched. Use user1 for testing.  (optional)
     * @return successful operation (status code 200)
     *         or error response (status code 400)
     *         or error response (status code 404)
     */
    @Operation(
        operationId = "getUser",
        summary = "Get user by parameters",
        tags = { "user" },
        responses = {
            @ApiResponse(responseCode = "200", description = "successful operation", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  User.class))),
            @ApiResponse(responseCode = "400", description = "error response", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  ResponseModel.class))),
            @ApiResponse(responseCode = "404", description = "error response", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  ResponseModel.class)))
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/searchUser",
        produces = { "application/json" }
    )
    default ResponseEntity<List<User>> getUser(
        @Parameter(name = "name", description = "The name that needs to be fetched. Use user1 for testing. ", schema = @Schema(description = "")) @Valid @RequestParam(value = "name", required = false) String name,
        @Parameter(name = "surname", description = "The surname that needs to be fetched. Use user1 for testing. ", schema = @Schema(description = "")) @Valid @RequestParam(value = "surname", required = false) String surname
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"password\" : \"password\", \"address\" : \"address\", \"surname\" : \"surname\", \"name\" : \"name\", \"id\" : 0, \"email\" : \"email\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}